requirements.txt
pyyaml==5.4.1
aiohttp==3.7.4
colorama==0.4.4
uvloop==0.15.2
click==8.0.1
plotly==5.1.0
pandas==1.2.5
dash==1.20.0